requirements.txt
Flask-SQLAlchemy
Flask
psycopg2-binary
gunicorn
flask-cors
firebase-admin
python-decouple
python-dotenv